HOCM-166 constrain the correctly generated recurrences
This fixes a bug wherein if the "earliest" constraint fell on an interval
at which a recurrence would not ordinarily occur (e.g. if the first of the
current month falling on a Tuesday during a weekly event on each Wednesday),
the generated recurrences would align with the "earliest" constraint interval
instead of aligning with the actual start date/time (e.g. weekly events
spanning multiple months would "snap" to the first day of the month.

This fixes it by first generating the recurrences and then clamping them,
instead of trying to be clever and setting RRULE DSTART to the "earliest"
constraint date.